+++ playwright-tests/tests/dashboard-2.test.js
@@ -38,77 +38,6 @@ test.describe('Dashboard Page Tests', () => {
         await expect(page.locator('.panel-header p')).toContainText('Test Panel');
     });
 
-    test('Verify dashboard loads correctly', async () => {
-        await expect(page.locator('#new-dashboard')).toBeVisible();
-        await expect(page.locator('.name-dashboard')).toBeVisible();
-        await expect(page.locator('#panel-container')).toBeVisible();
-        await expect(page.locator('.name-dashboard')).toContainText(uniqueName);
-    });
-
-    test('Edit panel', async () => {
-        const panelHeader = page.locator('.panel-header').first();
-        const editIcon = panelHeader.locator('img.panel-edit-li');
-
-        await panelHeader.hover();
-        await expect(editIcon).toBeVisible();
-        await editIcon.click();
-
-        const editPanel = page.locator('.panelEditor-container');
-        await expect(editPanel).toBeVisible();
-        await page.fill('#panEdit-nameChangeInput', 'Updated Panel Name');
-        await page.click('.panEdit-save');
-        await expect(page.locator('.panel-header p')).toContainText('Updated Panel Name');
-    });
-
-    test('View panel', async () => {
-        const panelHeader = page.locator('.panel-header').first();
-        const viewIcon = panelHeader.locator('img.panel-view-li');
-
-        await panelHeader.hover();
-        await expect(viewIcon).toBeVisible();
-        await viewIcon.click();
-
-        const viewPanel = page.locator('#viewPanel-container');
-        await expect(viewPanel).toBeVisible();
-        await expect(page.locator('#overview-button')).toHaveClass(/active/);
-
-        const editButton = page.locator('#edit-button');
-        await editButton.click();
-        await expect(page.locator('#edit-button')).toHaveClass(/active/);
-
-        const cancelButton = page.locator('.panelEditor-container #discard-btn');
-        await expect(cancelButton).toBeVisible();
-        await cancelButton.click();
-        await expect(page.locator('.panelEditor-container')).not.toBeVisible();
-        await expect(page.locator('#panel-container')).toBeVisible();
-    });
-
-    test('Delete panel', async () => {
-        const panelHeader = page.locator('.panel-header').first();
-        const optionsBtn = panelHeader.locator('#panel-options-btn');
-
-        const initialPanelCount = await page.locator('.panel').count();
-        await panelHeader.hover();
-        await expect(optionsBtn).toBeVisible();
-        await optionsBtn.click();
-
-        const dropdownMenu = page.locator('#panel-dropdown-modal');
-        await expect(dropdownMenu).toBeVisible();
-        const deleteOption = dropdownMenu.locator('.panel-remove-li');
-        await expect(deleteOption).toBeVisible();
-        await deleteOption.click();
-
-        const deleteConfirmDialog = page.locator('#panel-del-prompt');
-        await expect(deleteConfirmDialog).toBeVisible();
-        const confirmDeleteBtn = deleteConfirmDialog.locator('#delete-btn-panel');
-        await expect(confirmDeleteBtn).toBeVisible();
-        await confirmDeleteBtn.click();
-
-        await page.waitForTimeout(2000);
-        const finalPanelCount = await page.locator('.panel').count();
-        expect(finalPanelCount).toBe(initialPanelCount - 1);
-    });
-
     test('Change dashboard settings', async () => {
         await page.click('#db-settings-btn');
         await expect(page.locator('.dbSet-container')).toBeVisible();
